+++ lib/oidc/renewTokens.ts
@@ -11,15 +11,15 @@
  *
  */
 import { AuthSdkError } from '../errors';
-import { TokenParams, Tokens } from './types';
+import { RenewTokensParams, Tokens } from './types';
 import { getWithoutPrompt } from './getWithoutPrompt';
 import { renewTokensWithRefresh } from './renewTokensWithRefresh';
 import { getDefaultTokenParams } from './util';
 
 // If we have a refresh token, renew using that, otherwise getWithoutPrompt
 // eslint-disable-next-line complexity
-export async function renewTokens(sdk, options?: TokenParams): Promise<Tokens> {
-  const tokens = sdk.tokenManager.getTokensSync();
+export async function renewTokens(sdk, options?: RenewTokensParams): Promise<Tokens> {
+  const tokens = options?.tokens ?? sdk.tokenManager.getTokensSync();
   if (tokens.refreshToken) {
     return renewTokensWithRefresh(sdk, options || {}, tokens.refreshToken);
   }